backend/pyproject.toml
[project]
name = "vantage-backend"
version = "0.1.0"
description = "Vantage — Universal Goal Orchestrator Backend"
requires-python = ">=3.12"

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",

    # Azure SDKs
    "azure-cosmos>=4.7.0",
    "azure-storage-blob>=12.22.0",
    "azure-servicebus>=7.12.0",
    "azure-identity>=1.17.0",

    # Azure OpenAI + Semantic Kernel
    "semantic-kernel>=1.0.0",
    "openai>=1.40.0",

    # Microsoft Graph
    "msgraph-sdk>=1.5.0",

    # Parsing
    "pymupdf>=1.24.0",           # PDF parsing
    "beautifulsoup4>=4.12.0",    # HTML parsing
    "httpx>=0.27.0",             # HTTP client
    "yt-dlp>=2024.8.0",          # YouTube metadata

    # Observability
    "opentelemetry-api>=1.26.0",
    "opentelemetry-sdk>=1.26.0",
    "opentelemetry-instrumentation-fastapi>=0.47b0",
    "azure-monitor-opentelemetry-exporter>=1.0.0b27",

    # Utilities
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.5.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
